Fix image upload (dengan nama PK, News)

+++ app/Http/Controllers/Master/ProductKnowledgeController.php
@@ -156,7 +156,7 @@ class ProductKnowledgeController extends Controller
 
         // Upload file process
         ($request->upload_file != null) ?
-            $file_url = $this->fileUpload($request->upload_file, "productknowledge/".$this->getRandomPath()) : $file_url = "";
+            $file_url = $this->getUploadPathNameFile($request->upload_file, "productknowledge/".$this->getRandomPath(), $request->filename) : $file_url = "";
 
         if($request->upload_file != null) $request['file'] = $file_url;
 
@@ -216,6 +216,18 @@ class ProductKnowledgeController extends Controller
 
         // dd($request->all());
         $productKnowledge = ProductKnowledge::create($request->all());
+
+        if($request->upload_file != null){
+
+            /* Upload updated image */
+            $imagePath = explode('/', $productKnowledge->file);
+            $count = count($imagePath);
+            $imageFolder = "productknowledge/" . $imagePath[$count - 2];
+            $imageName = $imagePath[$count - 1];
+
+            $this->uploadFile($request->upload_file, $imageFolder, $imageName);
+
+        }
 
         return response()->json(['url' => url('/product-knowledge')]);
     }
@@ -309,25 +321,39 @@ class ProductKnowledgeController extends Controller
         }
 
         $productKnowledge = ProductKnowledge::find($id);
+        $oldFile = "";
 
-        if($request->upload_file != null) {
-            /* Delete File PDF */
-            if ($productKnowledge->file != "") {
-                $filePath = explode('/', $productKnowledge->file);
-                $count = count($filePath);
-                $folderpath = $filePath[$count - 2];
-                File::deleteDirectory(public_path() . "/file/productknowledge/" . $folderpath);
-            }
+        if($productKnowledge->file != null && $request->upload_file != null) {
+            /* Save old file path */
+            $oldFile = $productKnowledge->file;
         }
 
         // Upload file process
         ($request->upload_file != null) ?
-            $file_url = $this->fileUpload($request->upload_file, "productknowledge/".$this->getRandomPath()) : $file_url = "";
+            $file_url = $this->getUploadPathNameFile($request->upload_file, "productknowledge/".$this->getRandomPath(), $request->filename) : $file_url = "";
 
         if($request->upload_file != null) $request['file'] = $file_url;
 
         // Update data
     	$productKnowledge->update($request->all());
+
+    	if($productKnowledge->file != null && $request->upload_file != null) {
+
+            /* Delete File PDF */
+            $filePath = explode('/', $oldFile);
+            $count = count($filePath);
+            $folderpath = $filePath[$count - 2];
+            File::deleteDirectory(public_path() . "/file/productknowledge/" . $folderpath);
+
+            /* Upload updated file */
+            $filePath = explode('/', $productKnowledge->file);
+            $count = count($filePath);
+            $fileFolder = "productknowledge/" . $filePath[$count - 2];
+            $fileName = $filePath[$count - 1];
+
+            $this->uploadFile($request->upload_file, $fileFolder, $fileName);
+
+        }
 
         return response()->json(
             [
